requirements.txt
customtkinter>=5.2.0
cryptography>=42.0.0
Pillow>=10.0.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
pytest>=8.0.0
pyinstaller>=6.0.0
